refactor: replace AvatarProfileInput
