+++ src/pages/Profile/ProfileInfo/ProfileInfo.tsx
@@ -1,7 +1,7 @@
 import {FC} from "react";
 import {useParams} from "react-router-dom";
 import Grid from "@mui/material/Grid";
-import {AvatarProfileInput} from "@containers/AvatarProfileInput/AvatarProfileInput";
+import {AvatarProfileInput} from "@/pages/Profile/AvatarProfileInput";
 import {ProfileUserInfo} from "@/pages/Profile/ProfileUserInfo";
 import {ProfileForm} from "@/pages/Profile/ProfileForm/ProfileForm";
 import {ProgressSpinner} from "@view/ProgressSpinner/ProgressSpinner";
